+++ types/index.d.ts
@@ -221,7 +221,7 @@ declare namespace nock {
   interface Recorder {
     rec(options?: boolean | RecorderOptions): void
     clear(): void
-    play(): string[] | Definition[]
+    play(): Promise<string[] | Definition[]>
   }
 
   interface RecorderOptions {
@@ -252,17 +252,17 @@ declare namespace nock {
     fixtures: string
     setMode(mode: BackMode): void
 
-    (fixtureName: string, nockedFn: (nockDone: () => void) => void): void
+    (fixtureName: string, nockedFn: (nockDone: () => Promise<void>) => void): void
     (
       fixtureName: string,
       options: BackOptions,
-      nockedFn: (nockDone: () => void) => void,
+      nockedFn: (nockDone: () => Promise<void>) => void,
     ): void
     (
       fixtureName: string,
       options?: BackOptions,
     ): Promise<{
-      nockDone: () => void
+      nockDone: () => Promise<void>
       context: BackContext
     }>
   }