+++ CHANGELOG.md
@@ -4,3 +4,6 @@ Nock’s changelog can be found directly in the [GitHub release notes](https://g
 These are automatically created by [semantic-release](https://github.com/semantic-release/semantic-release) based on their [commit message conventions](https://semantic-release.gitbook.io/semantic-release#commit-message-format).
 
 Migration guides are available for major versions in the [migration guides directory](https://github.com/nock/nock/tree/main/migration_guides).
+
+Remove this before merge:
+1. Major - recorder.play and nockDone are async